--- lxl-web/src/routes/api/[[lang=lang]]/supersearch/qualifierTypes.test.ts
@@ -1,1 +1,0 @@
-// TODO
